fix: align and wrong folder
