+++ src/Parsers/Base.php
@@ -15,11 +15,13 @@ class Base {
 
 	public function set_settings( $settings ) {
 		$this->settings = (array) $settings;
+		
 		return $this;
 	}
 
 	public function parse_boolean_values() {
 		array_walk_recursive( $this->settings, [ $this, 'convert_string_to_boolean' ] );
+
 		return $this;
 	}
 
@@ -45,10 +47,14 @@ class Base {
 		}
 	}
 
+	protected function convert_number_to_string( &$value ) {
+		$value = (string) $value;
+	}
+
 	protected function remove_empty_values() {
 		foreach ( $this->settings as $key => $value ) {
 			// Remove empty string in an array.
-			$value = ! is_array( $value ) ? $value : array_filter( $value, function( $v ) {
+			$value = ! is_array( $value ) ? $value : array_filter( $value, function ($v) {
 				return $v !== '';
 			} );
 
@@ -59,6 +65,17 @@ class Base {
 
 			if ( $value === '' || $value === [] ) {
 				unset( $this->settings[ $key ] );
+			}
+
+			if ( $value === false ) {
+				if ( ! isset( $this->keep_false ) || ! is_array( $this->keep_false ) ) {
+					unset( $this->settings[ $key ] );
+					continue;
+				}
+
+				if ( ! in_array( $key, $this->keep_false ) ) {
+					unset( $this->settings[ $key ] );
+				}
 			}
 		}
 
@@ -136,7 +153,7 @@ class Base {
 				$condition = null;
 				continue;
 			}
-			$condition = [
+			$condition = [ 
 				$condition['name'],
 				$condition['operator'],
 				$condition['value'],
@@ -144,11 +161,26 @@ class Base {
 		}
 		$data['when'] = array_filter( $data['when'] );
 
+		$minimal_condition = $data;
+
+		// Remove relation if the value is AND
+		if ( 'and' === $data['relation'] ) {
+			unset( $minimal_condition['relation'] );
+		}
+
+		// If array contains only one item, convert it to a single array.
+		if ( count( $data['when'] ) === 1 ) {
+			$minimal_condition = reset( $data['when'] );
+			if ( $minimal_condition[1] === '=' ) {
+				$minimal_condition = [
+					$minimal_condition[0],
+					$minimal_condition[2],
+				];
+			}
+		}
+
 		if ( ! empty( $data['when'] ) ) {
-			$this->{$data['type']} = array(
-				'when'     => array_values( $data['when'] ),
-				'relation' => $data['relation'],
-			);
+			$this->{$data['type']} = $minimal_condition;
 		}
 
 		unset( $this->conditional_logic );
