+++ src/Parsers/MetaBox.php
@@ -6,8 +6,8 @@ use MetaBox\Support\Arr;
 class MetaBox extends Base {
 	protected $empty_keys = [ 'fields' ];
 	private $settings_parser;
-	private $validation = [
-		'rules'    => [],
+	private $validation = [ 
+		'rules' => [],
 		'messages' => [],
 	];
 
@@ -24,11 +24,8 @@ class MetaBox extends Base {
 
 		$this->remove_empty_values();
 
-		// Remove array keys again. Some methods like parse tabs change fields.
-		$this->fields = array_values( $this->fields );
-
 		$settings       = $this->settings_parser->get_settings();
-		$this->settings = array_merge( $settings, [ 'fields' => $this->fields ] );
+		
 		if ( $this->validation['rules'] ) {
 			if ( empty( $this->validation['messages'] ) ) {
 				unset( $this->validation['messages'] );
@@ -37,12 +34,17 @@ class MetaBox extends Base {
 		}
 
 		$this->settings = apply_filters( 'mbb_meta_box_settings', $this->settings );
+
+		// Remove array keys again. Some methods like parse tabs change fields.
+		$this->fields = array_values( $this->fields );
+		$this->settings = array_merge( [ 'fields' => $this->fields ], $settings );
 	}
 
 	private function parse_settings() {
-		$settings = [
+		$settings = [ 
 			'title' => $this->post_title,
-			'id'    => $this->post_name,
+			'id' => $this->post_name,
+			'modified' => $this->modified ?? time(),
 		];
 
 		if ( isset( $this->settings['settings'] ) ) {
